rename JwtFilter into JwtAuthenticateFilter, modify SecurityConfiguration
