+++ src/main/java/com/practice/demo/filter/JwtAuthenticateFilter.java
@@ -1,0 +1,46 @@
+package com.practice.demo.filter;
+
+import com.practice.demo.util.JwtUtil;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.filter.OncePerRequestFilter;
+import javax.servlet.*;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+
+public class JwtAuthenticateFilter extends OncePerRequestFilter {
+
+    public static final String LOGIN_PATH = "/api/login";
+    public static final String REGISTER_PATH = "/api/users/register";
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
+            throws ServletException, IOException {
+        String token = getTokenFromRequest(request);
+        String path = request.getRequestURI();
+
+        if (!path.equals(LOGIN_PATH) && !path.equals(REGISTER_PATH)) {
+            if (token != null && JwtUtil.validateToken(token)) {
+                Authentication authentication = JwtUtil.getAuthentication(token);
+                SecurityContextHolder.getContext().setAuthentication(authentication);
+            } else {
+                response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+                response.getWriter().write("Unauthorized - Invalid token");
+                return;
+            }
+        }
+
+        filterChain.doFilter(request, response);
+    }
+
+    private String getTokenFromRequest(HttpServletRequest request) {
+        String bearerToken = request.getHeader(JwtUtil.TOKEN_HEADER);
+        if (bearerToken != null && bearerToken.startsWith(JwtUtil.TOKEN_PREFIX)) {
+            return bearerToken.substring(7);
+        }
+        return null;
+    }
+}
+
